requirements.txt
PyQt6>=6.5.0
openpyxl>=3.1.0
onnxruntime>=1.16.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
py-cpuinfo>=9.0.0
